Fixed the "spinner"
Dug around in the webkit source and found the user-agent css for the
selector thingie on "number" type inputs.  Hid it.  Also moved the
validation icons back over again.

This feels fragile, and hasn't been tested on all of the browsers.
Another way to do this is to keep the input type "text" and specify a
pattern="([1-9]|1[012])" type thing, but then the browser doesn't
provide as helpful a validation message.