Added AM/PM boxes, fixed reveal close event
All times have an AM/PM box on them now.  As a side-effect, the icons
are now "right" instead of a fixed distance from the left of the text
box.

Also, the scrolling animation now fires on the event, "reveal:close"
instead of "click" on the "X" so it scrolls on any closing of the
dialog.  jQuery FTW.