Build a .htaccess file with a redirect for every path
This is part of the end-of-life process for this site.  As we move over
to drupal, we'd like to keep our old URL-space working.  This change
creates an .htaccess file that rewrites every html and pdf file to
"somewhere".  The default will be the new site root, but any page can be
over-written in the YAML file to somwhere else.